% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatQuad.R
\name{LQuad}
\alias{LQuad}
\title{Analyse eines lateinischen Quadrats}
\usage{
LQuad(
  file_path,
  sheet_name,
  block_col,
  column_col,
  treatment_col,
  response_cols
)
}
\arguments{
\item{file_path}{Pfad zur Excel-Datei}

\item{sheet_name}{Tabellenblatt}

\item{block_col}{Spaltenname für Block}

\item{column_col}{Spaltenname für Column}

\item{treatment_col}{Spaltenname für Treatment}

\item{response_cols}{Vektor mit Response-Spalten}
}
\value{
Liste mit ANOVA, LSD und Mittelwerten je Response
}
\description{
Führt eine ANOVA für ein lateinisches Quadrat durch und berechnet LSD sowie Mittelwerte.
}
